Rename complicated example to complex. Update documentation.
