+++ src-docs/src/views/partials.js
@@ -1,7 +1,7 @@
 export { default as KibanaChrome } from './kibana/kibana_chrome';
 export { default as KibanaHeader } from './header/header';
 export { default as TablePagination } from './pagination/customizable_pagination';
-export { default as ManagementSideNav } from './side_nav/side_nav_complicated';
+export { default as ManagementSideNav } from './side_nav/side_nav_complex';
 export { default as Table } from './table/table';
 export {
   default as ToastList,
